mm-video: vidc: Add support for picture order count
Add support for picture order count. By default PicOrderType
is set to 2 for H264 encode. Setting this parameter in the
sps/pps to 2 causes the decoder to return the first I-frame
immediately after decode, as opposed to holding onto it
until the next frame. Apps can configure it via OMX
Get/Set API.

Change-Id: I57c6287603747ef7e838e7017e2378a6c9e736af